refactor: toggle sidebar button
