+++ src/profile/components/ProfileHeader.tsx
@@ -2,7 +2,7 @@ import { Avatar, Heading, Text } from '@chakra-ui/react';
 
 export const ProfileHeader = () => {
   return (
-    <header className="fixed z-50 top-0 profile-header bg-primary-500 text-white flex h-[115px] justify-between items-center p-4">
+    <header className="fixed z-40 top-0 profile-header bg-primary-500 text-white flex h-[115px] justify-between items-center p-4">
       <div className="flex items-center gap-4">
         <Avatar size="xl" name="John Doe" src="" />
         <div className="flex flex-col">
